backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.1
supabase==2.5.0
httpx==0.27.0
PyJWT==2.10.1
requests==2.32.3
dockerfile-parse==2.0.1
langchain
langchain-google-genai==2.1.10
numpy>=1.26.0
